Move existing code and constants into TBA API utils module
