+++ scripts/frc/teams_counter/teams_counter.py
@@ -1,4 +1,3 @@
-import itertools
 import string
 import urllib.parse
 from collections.abc import Sequence
@@ -6,35 +5,15 @@ from collections.abc import Sequence
 import more_itertools
 import requests
 
+from scripts.frc.tba_api_utils import AUTH_KEY_HEADER, API_BASE_URL
 
-AUTH_KEY_HEADER = "X-TBA-Auth-Key"
-API_BASE_URL = "https://www.thebluealliance.com/api/v3/"
-TEAMS_SIMPLE_ENDPOINT = string.Template("teams/$page_number/simple")
+
 YEARS_PARTICIPATED_ENDPOINT = string.Template("team/$team_key/years_participated")
 TEAM_KEY_TEMPLATE = string.Template("frc$team_number")
 
 COUNTRY = "country"
 ISRAEL = "Israel"
 TEAM_NUMBER = "team_number"
-
-
-def get_all_teams(api_key: str) -> list[dict]:
-    """
-    :param api_key: TBA API auth key
-    :return: list of all teams
-    """
-    teams = []
-    for page_number in itertools.count():
-        url = urllib.parse.urljoin(
-            API_BASE_URL, TEAMS_SIMPLE_ENDPOINT.substitute(page_number=page_number)
-        )
-        response = requests.get(url, headers={AUTH_KEY_HEADER: api_key})
-        paged_teams = response.json()
-        if not paged_teams:
-            break
-        teams.extend(paged_teams)
-
-    return teams
 
 
 def get_israeli_teams(teams: list[dict]) -> list[dict]:
